Create universal file download logic
