+++ src/utility/url.ts
@@ -10,3 +10,19 @@ export const makeUrlAbsolute = (url?: string) => url !== undefined
   ? (!url.startsWith('http') ? `https://${url}` : url)
     .replace(/\/$/, '')
   : undefined;
+
+export const downloadFileFromBrowser = async (
+  url: string,
+  fileName: string,
+) => {
+  const blob = await fetch(url)
+    .then(response => response.blob());
+  const downloadUrl = window.URL.createObjectURL(blob);
+  const link = document.createElement('a');
+  link.href = downloadUrl;
+  link.download = fileName;
+  document.body.appendChild(link);
+  link.click();
+  document.body.removeChild(link);
+  window.URL.revokeObjectURL(downloadUrl);
+};
